# docs/swagger.yaml
basePath: /api
definitions:
...
  agent.EnvGetResponse:
    properties:
      key:
        example: DB_PASSWORD
        type: string
      value:
        example: secret123
        type: string
    type: object
  agent.EnvListResponse:
    properties:
      count:
        example: 3
        type: integer
      keys:
        example:
        - DB_HOST
        - DB_PASSWORD
        - API_KEY
        items:
          type: string
        type: array
    type: object
  agent.EnvSetRequest:
    properties:
      env:
        additionalProperties:
          type: string
        example:
          DB_HOST: localhost
          DB_PASSWORD: secret123
        type: object
    type: object
  agent.EnvSetResponse:
    properties:
      added:
        example:
        - DB_HOST
        - DB_PASSWORD
        items:
          type: string
        type: array
      updated:
        example:
        - API_KEY
        items:
          type: string
        type: array
    type: object
  agent.ErrorResponse:
    properties:
      error:
        example: workload not found
        type: string
    type: object
...
  agent.HostComposeProject:
    properties:
      containers:
        items:
          $ref: '#/definitions/agent.HostContainer'
        type: array
      managed_by_jetty:
        type: boolean
      name:
        type: string
      workload_name:
        description: populated when managed_by_jetty
        type: string
    type: object
  agent.HostContainer:
    properties:
      compose_project:
        type: string
      compose_service:
        type: string
      created_at:
        description: unix seconds
        type: integer
      id:
        description: short ID
        type: string
      image:
        description: image:tag
        type: string
      managed_by_jetty:
        type: boolean
      name:
        description: primary container name (without leading /)
        type: string
      ports:
        description: e.g. "0.0.0.0:80->80/tcp"
        items:
          type: string
        type: array
      state:
        description: '"running", "exited", "paused", ...'
        type: string
      status:
        description: human-readable e.g. "Up 5 hours"
        type: string
      workload_name:
        description: populated when managed_by_jetty
        type: string
    type: object
...
info:
  contact:
    name: Jetty Support
    url: https://github.com/ncwardell/jetty
  description: P2P Docker Compose orchestration with Cloudflare WARP mesh networking
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jetty API
  version: 0.0.x
paths:
...
  /env:
    get:
      description: Returns all stored environment variable keys (values are encrypted)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.EnvListResponse'
      summary: List environment variables
      tags:
      - env
    post:
      consumes:
      - application/json
      description: Stores encrypted environment variables (existing keys are overwritten)
      parameters:
      - description: Environment variables to set
        in: body
        name: env
        required: true
        schema:
          $ref: '#/definitions/agent.EnvSetRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.EnvSetResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
        "500":
          description: Encryption error
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Set environment variables
      tags:
      - env
  /env/{key}:
    delete:
      description: Removes an environment variable from storage
      parameters:
      - description: Environment variable key
        in: path
        name: key
        required: true
        type: string
      responses:
        "204":
          description: Variable deleted
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Delete environment variable
      tags:
      - env
    get:
      description: Returns the decrypted value of a specific environment variable
      parameters:
      - description: Environment variable key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.EnvGetResponse'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
        "500":
          description: Decryption error
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Get environment variable
      tags:
      - env
...
  /host/compose:
    get:
      description: Returns containers grouped by compose project. Standalone containers
        (no compose project) are grouped under an empty project name. Pass ?node=<id|name>
        to query a peer.
      parameters:
      - description: Peer ID or name; defaults to this node
        in: query
        name: node
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/agent.HostComposeProject'
            type: array
      summary: List Docker compose projects on this node
      tags:
      - host
  /host/containers:
    get:
      description: Returns every container on the host, including those not managed
        by Jetty. The agent's own container is excluded. Pass ?node=<id|name> to query
        a peer instead - useful when the dashboard is loaded via a shared Cloudflare
        tunnel and lands on whichever node Cloudflare picks.
      parameters:
      - description: Peer ID or name; defaults to this node
        in: query
        name: node
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/agent.HostContainer'
            type: array
      summary: List all Docker containers on this node
      tags:
      - host
...
schemes:
- https
- http
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: API key for authentication (JETTY_SECRET)
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"